prodexa/ml-service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
pydantic
python-dotenv
httpx